cruds

+++ odyssey/types.go
@@ -15,7 +15,7 @@ type Account struct {
 	FirstName string    `json:"firstName"`
 	Lastname  string    `json:"lastName"`
 	Number    int64     `json:"number"`
-	Balance   int64     `json:"balance"`
+	Balance   float64   `json:"balance"`
 	CreatedAt time.Time `json:"createdAt"`
 }
